Added helper function
Helper functions for firebase added
