+++ src/assets/css/_header.scss
@@ -81,7 +81,7 @@ nav{
         cursor: pointer;
     }
 }
-@media screen and (max-width: 1024px){
+@media screen and (max-width: 1524px){
     .nav-links{
         width: 40%;
     }
