+++ src/components/dashboard/dashboard.js
@@ -1,4 +1,4 @@
-import React, {useState} from 'react'
+import React from 'react'
 import Search from './search';
 import DashboardPanel from './dashboardPanel';
