+++ src/components/myAccount.js
@@ -1,14 +1,6 @@
-import React, {useState} from 'react'
-import  user  from '../assets/icons/user.svg';
+import React from 'react'
 
 function MyAccount() {
-
-    const [edit, setEdit] = useState(false);
-
-    const handleEdit = () => {
-        setEdit(!edit);
-        console.log(edit);
-    };
 
     return (
         <div className="container">
